requirements.txt
beautifulsoup4==4.11.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coverage==6.4.2
coveralls==3.3.1
docopt==0.6.2
flake8==5.0.4
flake8-quotes==3.3.1
idna==3.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-decouple==3.6
requests==2.28.1
requests-mock==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.11
